CL.py: adapted to newest version of CL library.
cdplayer.py: It is now also possible to give the CD's TOC as argument
to init().  The TOC is a string consisting of 2 digits giving the
number of tracks and then for each track the length of the track as
number of minutes and number of seconds (2 digits each).

cddb.py: Interface to the .cddb directory.  Usage is the same as for
cdplayer.py.


git-svn-id: http://svn.python.org/projects/python/trunk@3279 6015fed2-1504-0410-9fe1-9d1591cc4771