Much more rigorous testing -- we now try many times, varying in time of day, day of week, and season.
Fix the weekday predictions -- these seemed to be all bogus.  The new
predictions seem to correspond with strftime() on Solaris and IRIX, so
I believe they are correct.

Get rid of the test for non-standard format %C returning "the same as
date(1)".  This is hard to do reliably without opening a pipe to date,
and moreover, on IRIX 6.2, %C yields the Century.  So we use that
instead.  (We don't complain about this in non-verbose mode anyway.)


git-svn-id: http://svn.python.org/projects/python/trunk@7649 6015fed2-1504-0410-9fe1-9d1591cc4771