Mostly LaTeX style cleanup.
Use \newenvironment{envname} instead of \newcommand{\foo} &
\newcommand{\endfoo} (or \let\endfoo=...!) wherever reasonable.

Where {*desc} environment helper functions are not needed outside a
single environment definition, inline them.

Ensure that \seemodule{} and \seetext{} are only available within the
{seealso} environment.

Added "()" to the index entries for {funcdesc} and {cfuncdesc} environments.

Added {classdesc} environment which looks like the {funcdesc} environment
but doesn't add the parens in the index entry.


git-svn-id: http://svn.python.org/projects/python/trunk@9857 6015fed2-1504-0410-9fe1-9d1591cc4771