audio(7I) suggests that applications do the following to get the device and control pseudo-device:
- first look for the device filename in the environment variable
  AUDIODEV.

- if not found, use /dev/audio

- calculate the control device by tacking "ctl" onto the base device
  name.

We now do this.  Also, if the open fails, we call
PyErr_SetFromErrnoWithFilename() to give a more informative error
message.

Added a fileno() method to the audio object returned from open().
This returns the file descriptor which can be used by applications to
set up SIGPOLL notification, as per the manpage.


git-svn-id: http://svn.python.org/projects/python/trunk@11963 6015fed2-1504-0410-9fe1-9d1591cc4771