There are a few places which can raise DistutilsPlatformError; make sure it's imported!  ;)
Re-wrap the docstrings on get_python_inc() and get_python_lib() to be
closer to the "normal" Python style.  See GvR's "style guide" on the
essays page (http://www.python.org/doc/essays/).

There should never be a space between a function name and the '(' that
opens the argument list (see the style guide again).


git-svn-id: http://svn.python.org/projects/python/trunk@14631 6015fed2-1504-0410-9fe1-9d1591cc4771