Be even more careful about tables: explicitly translate the content of the table rather than letting the default processing take care of it.
The problem was that when multiple tables on a single output page used
different alignments, the specifier for the last table on the page was
used for all tables.  This is because LaTeX2HTML processes all the
enironments before processing other macros; internal state is
maintained for each table (column alignment!), which go overwritten as
additional tables were processed.  The saved state for the last table
processed was then used for all the tables on the page.

LaTeX2HTML is pain.


git-svn-id: http://svn.python.org/projects/python/branches/release152p1-patches@14686 6015fed2-1504-0410-9fe1-9d1591cc4771