Allow spaces in section names. Do not expose the __name__ when reporting the list of options available for a section since that is for internal use.
This closes SourceForge bug #115357.

Additionally, define InterpolationDepthError and MAX_INTERPOLATION_DEPTH.
The exception is raised by get*() when value interpolation cannot be
completed within the defined recursion limit.  The constant is only
informative; changing it will not affect the allowed depth.

Fix the exit from get() so that None is not returned if the depth is met
or exceeded; either return the value of raise InterpolationDepthError.


git-svn-id: http://svn.python.org/projects/python/trunk@17756 6015fed2-1504-0410-9fe1-9d1591cc4771