Hoepeful fix for SF bug #123924: Windows - using OpenSSL, problem with socket in httplib.py.
The bug reports that on Windows, you must pass sock._sock to the
socket.ssl() call.  But on Unix, you must pass sock itself.  (sock is
a wrapper on Windows but not on Unix; the ssl() call wants the real
socket object, not the wrapper.)

So we see if sock has an _sock attribute and if so, extract it.

Unfortunately, the submitter of the bug didn't confirm that this patch
works, so I'll just have to believe it (can't test it myself since I
don't have OpenSSL on Windows set up, and that's a nontrivial thing I
believe).


git-svn-id: http://svn.python.org/projects/python/trunk@18364 6015fed2-1504-0410-9fe1-9d1591cc4771