Fix core dump whenever PyList_Reverse() was called.
This fixes SF bug #132008, reported by Warren J. Hack.

I'm checking this into the 1.6 branch of the CVS tree first.  I'm
assigning the copyright for this patch (and this patch only) to CNRI,
at their specific request.  In order to be able to release 1.6.1 with
a GPL-compatible license, CNRI wants to make at least one change to
the sources.  This is it. :-)


git-svn-id: http://svn.python.org/projects/python/branches/cnri-16-start@19442 6015fed2-1504-0410-9fe1-9d1591cc4771