Use PyGeneric_{Get,Set}Attr.  We still need wrappers for the peculiar additional constraints on the type of some special attributes and the tests for restricted execution.
(Note that the restricted access test in func_getattro doesn't make
much sense.  In restricted mode, it disallows getting attributes whose
name *doesn't* start with underscore -- this forbids access to most
user attributes, but allows (writable!) access to the __dict__
attribute.  This was probably never reviewed when user attributes were
added.  Unclear what should be done, so for now I'm leaving it in.)


git-svn-id: http://svn.python.org/projects/python/branches/descr-branch@20763 6015fed2-1504-0410-9fe1-9d1591cc4771