INCREF the type object in type_call() when the type object is allocated on the heap.  This removes the need for subtype_construct() (which existed *just* to INCREF self->ob_type).  We still need subtype_dealloc() to do the matching DECREF, because the base type's dealloc won't do this.
I don't want to require all types to INCREF the type object for each
object they own, but we do need this for type objects allocated on the
heap.  Doing the INCREF in subtype_construct was too late, because if
subtype_construct() failed, type_call() would DECREF the partially
initialized object, which would invoke subtype_dealloc(), which would
incorrect DECREF'ed the type object.  The only safe thing to do is to
INCREF the type in type_call(), which this patch does.

Also correct a bug in subtype_dealloc(), which would call
base->tp_dealloc, even though it intended to call f, which it had
calculated as the first base->base->...->tp_dealloc found that wasn't
subtype_dealloc.


git-svn-id: http://svn.python.org/projects/python/branches/descr-branch@20858 6015fed2-1504-0410-9fe1-9d1591cc4771