PyDescr_IsData(): change in specification.  If the argument is not a PyDescrObject, this returns true if the argument's type supports the tp_descr_set slot.  This allows us to created computed attributes from a pair of get/set functions (see test_descr.py example).
This almost works for classic classes too: there, it can be used for
read-only computed attributes, but it can't be used to trap setattr
unless we change instance setattr to look in the class dict first.
Since that's expensive (a chain of dict lookups) we don't do this.

Why is it not expensive for new-style classes?  Because the class
lookup is faster: compare _PyType_Lookup() to class_lookup().


git-svn-id: http://svn.python.org/projects/python/branches/descr-branch@21454 6015fed2-1504-0410-9fe1-9d1591cc4771