THought I'd start the merge with something "obviously easy".  Heh. One thing leads to another, and test_unicode.py fails now.  But it's due to this bug:
>>> "%*s" % (5, "abc")
'  abc'
>>> "%*s" % (5, u"abc")
Traceback (most recent call last):
  File "<stdin>", line 1, in ?
TypeError: * wants int
>>>

This *form* of test (a %*s format) didn't used to exist in test_unicode.py,
so my best guess is that this is a legit bug in the descr-branch.  If so,
no reason not to check this in; perhaps the true error is somewhere in
still untouched Unicode-specific merges, though.


git-svn-id: http://svn.python.org/projects/python/branches/descr-branch@21504 6015fed2-1504-0410-9fe1-9d1591cc4771