Convert to string methods.
For the real document element, make sure the prolog is migrated into
the document element so it isn't left stranded.

Make fixup_trailing_whitespace() whitespace do what was really intended.

Add the *desc environments used in the C API manual to the list of
things that can exist at the paragraph level so they don't get wrapped
in <para>...</para>.


git-svn-id: http://svn.python.org/projects/python/trunk@23413 6015fed2-1504-0410-9fe1-9d1591cc4771