backport gvanrossum's checkin of     revision 2.129 of typeobject.c
"Fix" for SF bug #520644: __slots__ are not pickled.

As promised in my response to the bug report, I'm not really fixing
it; in fact, one could argule over what the proper fix should do.
Instead, I'm adding a little magic that raises TypeError if you try to
pickle an instance of a class that has __slots__ but doesn't define or
override __getstate__.  This is done by adding a bozo __getstate__
that always raises TypeError.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@25693 6015fed2-1504-0410-9fe1-9d1591cc4771