Fix for SF bug 528132 (Armin Rigo): classmethod().__get__() segfault
The proper fix is not quite what was submitted; it's really better to
take the class of the object passed rather than calling PyMethod_New
with NULL pointer args, because that can then cause other core dumps
later.

I also added a testcase for the fix to classmethods() in test_descr.py.

I'll apply this to 2.3 too.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@25727 6015fed2-1504-0410-9fe1-9d1591cc4771