test_no_parts_in_a_multipart(): A test for the layout of a multipart/mixed message with no attachments.
test_one_part_in_a_multipart(): A test for the layout of a
multipart/mixed message with a single attachment.

test_seq_parts_in_a_multipart(): A test for the layout of a
multipart/mixed message with a single attachment that happens to be a
sequence of length one.

These tests ensure no regressions on the fix for SF bug #531966.

I will merge these into the standalone email package and Python 2.3
trunk separately.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@25813 6015fed2-1504-0410-9fe1-9d1591cc4771