backport jackjansen's checkin of     revision 1.211 of socketmodule.c
Due to interaction between the MSL C library and the GUSI I/O library I can get reads from sockets to work consistently either for unbuffered binary files or for buffered binary files, but not for both:-(

The workaround is to force socket.makefile() to disable buffering for binary files.

Fixes bug 534625. 2.2.1 candidate.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@25913 6015fed2-1504-0410-9fe1-9d1591cc4771