Backport rev 2.143 (note: some earlier bugfix candidates still TBD).
SF bug 542984.

Change type_get_doc (the get function for __doc__) to look in tp_dict
more often, and if it finds a descriptor in tp_dict, to call it (with
a NULL instance).  This means you can add a __doc__ descriptor to a
new-style class that returns instance docs when called on an instance,
and class docs when called on a class -- or the same docs in either
case, but lazily computed.

I'll also check this into the 2.2 maintenance branch.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@26481 6015fed2-1504-0410-9fe1-9d1591cc4771