Closes: #556025 seg fault when doing list(xrange(1e9))
A MemoryError is now raised when the list cannot be created.
There is a test, but as the comment says, it really only
works for 32 bit systems.  I don't know how to improve
the test for other systems (ie, 64 bit or systems
where the data size != addressable size,
e.g. 64 bit data, but 48 bit addressable memory)


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@26884 6015fed2-1504-0410-9fe1-9d1591cc4771