Backport to 2.2.x:
SF bug 533625 (Armin Rigo). rexec: potential security hole

If a rexec instance allows writing in the current directory (a common
thing to do), there's a way to execute bogus bytecode.  Fix this by
not allowing imports from .pyc files (in a way that allows a site to
configure things so that .pyc files *are* allowed, if writing is not
allowed).


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@26968 6015fed2-1504-0410-9fe1-9d1591cc4771