Added character data buffering to pyexpat parser objects.
Setting the buffer_text attribute to true causes the parser to collect
character data, waiting as long as possible to report it to the Python
callback.  This can save an enormous number of callbacks from C to
Python, which can be a substantial performance improvement.

buffer_text defaults to false.


git-svn-id: http://svn.python.org/projects/python/trunk@27408 6015fed2-1504-0410-9fe1-9d1591cc4771