Change the macros used for visitor and finish symtable visitor.
This code seems to be a clear improvement over the old compile.c
version.  It was wriiten in about a quarter of the time.  It is
shorter (350 lines vs. 510 lines).  It's independent of Grammar.  And
it in't filled with magic constants.

XXX A uture revision should extract out the navigation part of the
code and make it generic.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@27504 6015fed2-1504-0410-9fe1-9d1591cc4771