Backport patches for versions 1.33, 1.35, and 1.36 from the trunk:
revision 1.36:
If PyXML is installed, there is no Node.allnodes, so that portion of
the test should be skipped if that's the case.

revision 1.35:
Remove duplicate checks of the Node.allnodes variable.

revision 1.33:
Follow PyXML: Remove all prints from successful tests.  This means we can
also drop the output file.


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@28886 6015fed2-1504-0410-9fe1-9d1591cc4771