(py-parse-state-re): Remove the "if" from the regular expression. This fixes an indentation bug reported by Jeremy when seeing multiple list comprehensions like so:
    [x for x in seq
     if blah(x)]

    # ...

    [y for y in seq
     if blah(y)]

The reason this broke is because this regexp caused the "find a safe
parsing start location higher up in the file" test to erroneously find
the if in the listcomp.  I think the other keywords in this regexp are
fine and good enough.

After a weekend of testing, I can't find any adverse effects.


git-svn-id: http://svn.python.org/projects/python/trunk@29346 6015fed2-1504-0410-9fe1-9d1591cc4771