Some help for SF 614770:  MSVC 7.0 compiler support
This changes sys.version under Microsoft builds to include the MS compiler
version number (_MSC_VER).  Since VC 6 and VC 7 are apparently
incompatible, and both can be installed on a single box, distutils needs
some way to figure out which version of MSVC a given Python was compiled
under.

As also suggested by MvL, got rid of #ifdef'ery for the defunct _M_ALPHA
target.

Bugfix candidate?  Hard to say.  As far as I'm concerned, VC 7 wasn't
a supported platform in the 2.2 line.  If somebody thinks it should be,
they can do the work.


git-svn-id: http://svn.python.org/projects/python/trunk@29626 6015fed2-1504-0410-9fe1-9d1591cc4771