Backport Patch 659834 checked in by GvR on 2002/12/30 16:25:38
Check for readline 2.2 features.  This should make it possible to
compile readline.c again with GNU readline versions 2.0 or 2.1; this
ability was removed in readline.c rev. 2.49.  Apparently the older
versions are still in widespread deployment on older Solaris
installations.  With an older readline, completion behavior is subtly
different (a space is always added).


git-svn-id: http://svn.python.org/projects/python/branches/release22-maint@32053 6015fed2-1504-0410-9fe1-9d1591cc4771