Backport Tim's fix to test_winsound.py
PlaySoundTest.test_alias_fallback():  Disabled this test, and explained
why in a new comment.  My home Win98SE box is one of the "real systems"
alluded to (my system "default sound" appears to have vanished sometime
in the last month, that's certainly not a Python bug, and the MS
PlaySound docs are correct in their explanation of what happens then).


git-svn-id: http://svn.python.org/projects/python/branches/release23-maint@34223 6015fed2-1504-0410-9fe1-9d1591cc4771