Backport to 2.3:
If a file is opened with an explicit buffer size >= 1, repeated
close() calls would attempt to free() the buffer already free()ed on
the first close().     [bug introduced with patch #788249]

Making sure that the buffer is free()ed in file object deallocation is
a belt-n-braces bit of insurance against a memory leak.


git-svn-id: http://svn.python.org/projects/python/branches/release23-maint@35642 6015fed2-1504-0410-9fe1-9d1591cc4771