Fix several symbol table/compiler problems uncovered by test_scope.
The symbol table wasn't accounting for nested class blocks with local
variables that shadow free variables of methods.  In this case, the
name is added to co_freevars and is treated as a local variable.  The
generated code calls LOAD_CLOSURE to pass the cell for the free
variable from the enclosing scope through to the methods.

XXX Use the DEF_FREE_CLASS flag to mark classes where the same name is
local and free.  This is a bit of a hack, because it's a scoping issue
and scope decisions are stored elsewhere.  But the bits used for
scoping decisions are densely packed (1, 2, 3, 4, and 5 all have
meanings) instead of being a bitfield.  Perhaps we should revise that
and use one bit for each scope; then this special case would just have
two scope bits set.

Modify dictbytype() to take a scope and a flag, so that free variables
can be extracted correctly.

Cell variables were being created in class blocks, which doesn't make
any sense.  A class can never have a cell variable, because it's
bindings aren't visible to nested scopes.

Global variables weren't being handled properly.  Extend
analyze_block() to thread through a set containing the current defined
global variables.  A bizarre feature of the scoping rules is that a
global statement in an enclosing scope affects a nested scope.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@35664 6015fed2-1504-0410-9fe1-9d1591cc4771