Recompiled the binary wininst.exe after this change:
When loading the Python dll to run the postinstall script, try to load
it from the install directory (as reported by the registry) in case it
is not found on the default Loadlibrary search path.

Fixes SF 935091: bdist_winist post-install script fails on non-admin Python

Will port to the trunk later.


git-svn-id: http://svn.python.org/projects/python/branches/release23-maint@35704 6015fed2-1504-0410-9fe1-9d1591cc4771