Simplify hairy code to initialize cell variables.
The old code depended on the order in which the old compiler put names
into co_cellvars.  The order is different with the new compiler, which
caused a bunch of test failures (e.g. test_userdict) when an argument
was also a cell variable.

In comment, note that it might be a good idea for the compiler to
generate names in a certain order.  On the other hand, it's hard to
imagine a practical case where this would make a difference.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@35729 6015fed2-1504-0410-9fe1-9d1591cc4771