Fix code generation for extended slices.
The AST should probably be changed for slices, because an
ExtendedSlice is sufficiently unlike any of the other cases.  In
particular, an ExtSlice can contain any of the other slice types, and
it is the only slice type that can do this.

Also, fix stack depth calculation for STORE+n opcodes.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@35734 6015fed2-1504-0410-9fe1-9d1591cc4771