Repair failure of test_difflib.  difflib has a nasty one:
        >>> for line in results: print repr(line)
        ...
        [results go here]

That managed to break the intent (of doctest's parsing) that an
example would end with a newline iff it was a multi-line example.
That example really isn't multiline, but ended up with a trailing
newline anyway.  Edward's new code asserted this invariant, which
the old difflib code did not do.  So fixed the parsing to get rid
of the trailing newline in this case.

But then it turned out that compile(... "single") doesn't accept
a "block opener: body" statement *without* a trailing newline,
raising SyntaxError if you try.  So changed compilation to always
append a newline.

test_difflib passes now.


git-svn-id: http://svn.python.org/projects/python/branches/tim-doctest-branch@36765 6015fed2-1504-0410-9fe1-9d1591cc4771