Backport of fixes for #887242 and #1097739:
If $MACOSX_DEPLOYMENT_TARGET is set, and >= 10.3, during configure we
setup extensions to link with dynamic lookup. We also record the
value in the Makefile, and distutils uses the same value to build extension
modules. If MACOSX_DEPLOYMENT_TARGET is not set it defaults to the
current OSX version.

If we cannot use -undefined dynamic_lookup (such as on 10.2 or earlier)
we link extensions directly against the dynamic library in the framework in
stead of against the framework. This will fix building extensions for 2.3
after 2.4 has been installed too.


git-svn-id: http://svn.python.org/projects/python/branches/release23-maint@38269 6015fed2-1504-0410-9fe1-9d1591cc4771