Running a bdist_wininst installer, built with Python 2.3, installing for Python 2.4 caused a segfault when post_install_script was used.
The reason was that the file handle passed to PyRun_SimpleFile() was
created with MSVCRT.DLL, but Python 2.4 uses MSVCR71.DLL.

So, I replaced PyRun_SimpleFile() with PyRun_SimpleString().  The
segfault is gone, but the output of the postinstall script doesn't
show up, because still freopen() from MSVCRT is used.


git-svn-id: http://svn.python.org/projects/python/branches/release24-maint@38415 6015fed2-1504-0410-9fe1-9d1591cc4771