Backport checkin: Fix for SF bug #1175396: readline() will now read one more character, if the last character read is "\r" (and size is None, i.e. we're allowed to call read() multiple times), so that we can return the correct line ending (this additional character might be a "\n").
If the stream is temporarily exhausted, we might return the wrong line ending
(if the last character read is "\r" and the next one (after the byte stream
provides more data) is "\n", but at least the atcr member ensures that we
get the correct number of lines (i.e. this "\n" will not be treated as
another line ending).


git-svn-id: http://svn.python.org/projects/python/branches/release24-maint@38772 6015fed2-1504-0410-9fe1-9d1591cc4771