test_default_encoding_issues():  Fully restore sys.setdefaultencoding.
test_site often failed under "regrtest.py -r", because this xmlrpc test
left sys with a setdefaultencoding attribute, but loading site.py removes
that attribute and test_site.py verifies the attribute is gone.  Changed
this test to get rid of sys.setdefaultencoding if it didn't exist when
this test started.

Don't know whether this is a bugfix (backport) candidate.


git-svn-id: http://svn.python.org/projects/python/trunk@38786 6015fed2-1504-0410-9fe1-9d1591cc4771