Fix two dotted import problems uncovered by zope test suite
The symbol table did not handle dots in import names.  If you imported
"a.b.c", it would put "a.b.c" in the symbol table instead of "a".

The compiler did not generate the correct code for "import a.b.c as
d".  It bound a to d instead of c to d.

Add very simple tests cases to test_import to cover these failures.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@39726 6015fed2-1504-0410-9fe1-9d1591cc4771