Fix handling of testlist_gexp.
Created a single function that handles the several possible testlist
variants, with a flag indicating whether a gexp is possible.  Note
that I don't understand why that Grammar allows testlist_gexp in some
places but not others.

Address several other XXX comments, including calls to free.  Reindent
a few functions to four spaces so that they can be read.


git-svn-id: http://svn.python.org/projects/python/branches/ast-branch@39738 6015fed2-1504-0410-9fe1-9d1591cc4771