Implement memory management for AST using an arena API.
The key idea is to allocate all AST notes from a single arena that is
freed when the compilation finishes (success or failure).  The change
is pervasive:  Every AST constructor now takes an arena as an
additional argument.  Also, every call related to freeing AST memory
is gone.

The actual implementation of the arena is a toy.  It uses plain old
malloc and free and stores all the pointers so allocated in a linked
list.

The arena also keeps track of PyObjects allocated by the AST and
DECREFs each of them when the arena is freed.  This part of the
implementation would probably stay, although Fredrik Lundh noted that
a PyList would server as a fine container for the pointers.

This set of changes only modifies the compiler in two places, where
Attributes are created.  In theory, the entire compiler could use an
AST interface.



git-svn-id: http://svn.python.org/projects/python/branches/ast-arena@41604 6015fed2-1504-0410-9fe1-9d1591cc4771