Repair the comments for Py_ADDRESS_IN_RANGE.
Alas, the patch introduced a hole in this logic, making it
possible for Py_ADDRESS_IN_RANGE(P) to believe that an address
P is controlled by obmalloc when it in fact is not.  Not yet
sure how to repair this, so added an XXX block explaining the
new mess (this can't be ignored:  it's a disaster if
Py_ADDRESS_IN_RANGE can ever be wrong).


git-svn-id: http://svn.python.org/projects/python/branches/tim-obmalloc@42658 6015fed2-1504-0410-9fe1-9d1591cc4771