Merge rev 42904 from the trunk.
_hotshot hotshot_profiler():  If write_header() returned
an error code, this let `self` leak.  This is a disaster
on Windows, since `self` already points to a newly-opened
file object, and it was impossible for Python code to
close the thing since the only reference to it was in a
blob of leaked C memory.

test_hotshot test_bad_sys_path():  This new test provoked
the C bug above.  This test passed, but left an open
"@test" file behind, which caused a massive cascade of
bogus test failures in later, unrelated tests on Windows.
Changed the test code to remove the @test file it leaves
behind, which relies on the change above to close that
file first.


git-svn-id: http://svn.python.org/projects/python/branches/release24-maint@42966 6015fed2-1504-0410-9fe1-9d1591cc4771