Merge rev 43734 from the trunk.
test_timeout():  This test was added during Bug Day, but disabled
soon after because the gmail address it connects to started timing
out on all the buildbot slaves.  Rewrote the test to produce a
warning message (instead of failing) when the address times out.

Also removed the special case for Windows -- this test started to
work on Windows as soon as bug 1462352 was fixed.


git-svn-id: http://svn.python.org/projects/python/branches/release24-maint@43735 6015fed2-1504-0410-9fe1-9d1591cc4771