Patch #1538606, Patch to fix __index__() clipping.
I modified this patch some by fixing style, some error checking, and adding
XXX comments.  This patch requires review and some changes are to be expected.
I'm checking in now to get the greatest possible review and establish a
baseline for moving forward.  I don't want this to hold up release if possible.



git-svn-id: http://svn.python.org/projects/python/trunk@51236 6015fed2-1504-0410-9fe1-9d1591cc4771