Remove tp_init from 'file'.  Left tp_new so as to continue to allow subclassing, although mostly useless since without tp_init you cannot actually open any file.
Added PyFile_UnsafeOpen() and set built-in open to that.  Also changed bz2
module to use it as well.

Still need to make C API calls go through built-in open() so that if it is
overridden that is used instead of blindly opening any file.  Will most likely
require a new set of APIs for that.


git-svn-id: http://svn.python.org/projects/python/branches/bcannon-objcap@51656 6015fed2-1504-0410-9fe1-9d1591cc4771