Fixed bug #1531862: Do not close standard file descriptors in subprocess.
Let's try that once more. Buildbots were broken last time, but probably
because tests were sending data to stderr for testing it (sending to a
file doesn't touch the problem).

The fix is still the same, but tests were reduced (removing tests to
be able to fix something is weird, but oh well).


git-svn-id: http://svn.python.org/projects/python/branches/release25-maint@51793 6015fed2-1504-0410-9fe1-9d1591cc4771