Backport of 52811: Bug #1599782: Fix segfault on bsddb.db.DB().type().
The problem is that _DB_get_type() can't be called without the GIL
because it calls a bunch of PyErr_* APIs when an error occurs.
There were no other cases in this file that it was called without the GIL.
Removing the BEGIN/END THREAD around _DB_get_type() made everything work.


git-svn-id: http://svn.python.org/projects/python/branches/release25-maint@52812 6015fed2-1504-0410-9fe1-9d1591cc4771