[Bug #1613651] Document socket.recv_into, socket.recvfrom_into
Also, the text for recvfrom told you to read recv() for an explanation of the
'flags' argument, but recv() just pointed you at the man page.  Copied the 
man-page text to recvfrom(), recvfrom_into, recv_into to avoid the pointless
redirection.

I don't have LaTeX on this machine; hope my markup is OK.


git-svn-id: http://svn.python.org/projects/python/trunk@53066 6015fed2-1504-0410-9fe1-9d1591cc4771