Patch from David Watson (the unified2-module.diff attached to bug #1599254).
Single-file mailboxes used to write the new mailbox contents into a
temporary file which is then renamed over the original. Unfortunately,
if another program tries to deliver messages while mailbox.py is
working, and uses only fcntl() locking, it will have the old file open
and be blocked waiting for the lock to become available. Once
mailbox.py has replaced the old file and closed it, making the lock
available, the other program will write its messages into the
now-deleted "old" file, consigning them to oblivion.

The fix is to copy new_file's contents back to the original file
instead of renaming a new file. 

If file.truncate() is available, the mailbox is then truncated to
size. Otherwise, if truncation is required, we just re-open the
destination file anew (and still have the original locking problem).
So this patch still has a race condition on platforms where there's no
file.truncate(), but it's not clear that any such platforms actually
exist at this time.  (AFAIK, the OS/2 VAC++ port lacks it, but Andrew
MacIntyre isn't maintaining it; the EMX OS/2 is the actively developed
one.) Is it worth keeping that branch of code?


git-svn-id: http://svn.python.org/projects/python/branches/amk-mailbox@54598 6015fed2-1504-0410-9fe1-9d1591cc4771