All the operations that already existed pass ok the new tests (except power).  For this, I fixed some bugs, reordered some code (without changing functionality), and made only one important change: Added a P value to the valid exponents, because we needed a different diagnostic information.
So far, for diagnostic information we used the Decimal._int digits,
because that was all that was needed. I think it's more clear to 
leave there the payload digits, and not make that structure more 
complex, and create the new exponent.

P is from Phantom signal, because for everybody asking 
self._isnan(), it will return 1 (quiet NaN).  But if you want to 
look at it in more detail, you will now that in a past life, it was 
signaled.



git-svn-id: http://svn.python.org/projects/python/branches/decimal-branch@54893 6015fed2-1504-0410-9fe1-9d1591cc4771