Recommit r54805:
Add code to read from master_fd in the parent, breaking when we get an OSError
(EIO can occur on Linux) or there's no more data to read.  Without this,
test_pty.py can hang on the waitpid() because the child is blocking on the
stdout write.  This will definitely happen on Mac OS X and could potentially
happen on other platforms.  See the comment for details.


git-svn-id: http://svn.python.org/projects/python/branches/release25-maint@54924 6015fed2-1504-0410-9fe1-9d1591cc4771