On Windows, when import fails to load a dll module, the message says "error code 193" instead of a more informative text.
It turns out that FormatMessage needs additional parameters for some error codes.
For example: 193 means "%1 is not a valid Win32 application".
Since it is impossible to know which parameter to pass, we use 
FORMAT_MESSAGE_IGNORE_INSERTS to get the raw message, which is still better
than the number.

Also use the Unicode version of the API, to deal with accented letters.



git-svn-id: http://svn.python.org/projects/python/branches/py3k@59682 6015fed2-1504-0410-9fe1-9d1591cc4771