#Issue3088 in-progress: Race condition with instances of classes derived from threading.local:
When a thread touches such an object for the first time, a new thread-local __dict__ is created,
and the __init__ method is run.
But a thread switch can occur here; if the other thread touches the same object, it installs another
__dict__; when the first thread resumes, it updates the dictionary of the second...

This is the deep cause of the failures in test_multiprocessing involving "managers" objects.

Backport of r64601.


git-svn-id: http://svn.python.org/projects/python/branches/release25-maint@64641 6015fed2-1504-0410-9fe1-9d1591cc4771