Modify format_float_short() to compute how much memory to allocate. This removes the last hard-coded buffer in this file.
Mark: Please review this, I'm sure it could be simplified. In
particular, min_digits1 feels hackish. Also, I'm assuming max 3 digits
for the exponent, which might not be correct. Maybe we should enforce
that the abs(exponent) < 1000?

Still need to detect errors in return value from_Py_dg_dtoa and handle
that correctly.



git-svn-id: http://svn.python.org/projects/python/branches/py3k-short-float-repr@71330 6015fed2-1504-0410-9fe1-9d1591cc4771