#6323: pdb doesn't deal well with SyntaxErrors.
It seems necessary to keep two layers of 'exec' (one in Bdb.run, one in Pdb._runscript);
this allows the tracing function to be active when the inner 'exec' runs
and tries to compile the real code.

This partially revert r58127, the net effet of the two changes is to replace
"exec('%s')" with "exec(%r)".


git-svn-id: http://svn.python.org/projects/python/branches/py3k@73918 6015fed2-1504-0410-9fe1-9d1591cc4771