Issue #7472: remove unused code from email.encoders.encode_7or8bit.
Yukihiro Nakadaira noticed a typo in encode_7or8bit that was trying
to special case iso-2022 codecs.  It turns out that the code in
question is never used, because whereas it was designed to trigger
if the payload encoding was eight bit but its output encoding was
7 bit, in practice the payload is always converted to the 7bit 
encoding before encode_7or8bit is called.  Patch by Shawat Anand.


git-svn-id: http://svn.python.org/projects/python/trunk@80800 6015fed2-1504-0410-9fe1-9d1591cc4771