#8575 - Update and reorganize some _winreg contents.
I've removed the hopeful note about a future higher-level module since
it's been in there for quite a long time and nothing of the sort has 
come up. There are a few places where markup was added to cross-reference
other sections, and many of the external links have been removed and now
point to newly created sections containing previously undocumented
information.

The Value Types section was created and it's contents were taken from
a function-specific area, since it applies to more than just that
function. It fits in better with the other newly documented constants.


git-svn-id: http://svn.python.org/projects/python/trunk@81088 6015fed2-1504-0410-9fe1-9d1591cc4771