Without this patch the value of sysconfig.get_config_var('LDSHARED') is wrong when PY_LDFLAGS is not empty.
The bug was caused by LDSHARED getting expanded *before* sysconfig
renamed PY_LDSHARED (and simular values) to names without a PY_
prefix.

The patch tries to maintain the intended behaviour of allowing users
to set LDFLAGS in the environment and have that affect the build.

Without this patch a universal build on OSX cannot build universal
(fat binary) extensions.



git-svn-id: http://svn.python.org/projects/python/branches/py3k@82995 6015fed2-1504-0410-9fe1-9d1591cc4771