Backport of 83120 (#9032)
XML-RPC client: Transport.request() retries on EPIPE error

The EPIPE error occurs when the server closes the socket and the client sends a
"big" XML-RPC request (I don't know exactly the size threshold).

request() just have to ignore the error because single_request() closes the
socket on error, and so the next call to single_request() will open a new
socket.

Remove also a comment in the HTTP client because it's now wrong: see r70643 and
issue #5542.


git-svn-id: http://svn.python.org/projects/python/branches/release27-maint@83123 6015fed2-1504-0410-9fe1-9d1591cc4771