Revert 83784 adding functools.lru_cache() to the re module.
The problem is that the re module is imported by sysconfig
and re needs functools which uses collections.OrderedDict()
but the _collectionsmodule.c code is not yet constructed
at this point in the build.

The likely best solution will be to include _collections
as part of the static build before the rest of the
boot-strapping.




git-svn-id: http://svn.python.org/projects/python/branches/py3k@83875 6015fed2-1504-0410-9fe1-9d1591cc4771