Issue #9065: no longer use "root" as the default for the uname and gname field.
If tarfile creates a new archive and adds a file with a
uid/gid that doesn't have a corresponding name on the
system (e.g. because the user/group account was deleted) it
uses the empty string in the uname/gname field now instead
of "root". Using "root" as the default was a bad idea
because on extraction the uname/gname fields are supposed
to override the uid/gid fields. So, all archive members
with nameless uids/gids belonged to the root user after
extraction.



git-svn-id: http://svn.python.org/projects/python/branches/py3k@85211 6015fed2-1504-0410-9fe1-9d1591cc4771