issue 10683 When the solution is converted to Visual Studio 2010, the command line to invoke make_buildinfo changes from: $(SolutionDir)make_buildinfo.exe" Debug "$(IntDir)\" to $(SolutionDir)make_buildinfo.exe" Debug "$(IntDir)" If the final backslash is omitted, the backslash in IntDir will escape the quote, thus passing the quote in as part of the path name.
This solution is a hack-fix to that problem by skipping any trailing quote from the path name.  It works as long as we don't need any additional arguments to make_buildinfo.exe.  This will help all those sould that are going to run this project through the visual studio autoconverter and get the same error.

git-svn-id: http://svn.python.org/projects/python/branches/py3k@87205 6015fed2-1504-0410-9fe1-9d1591cc4771