#4871: check that zipfile password is bytes, and give useful error message.
Previously passing a string in as the password would fail either with
an assertion error or a TypeError with a confusing error message.
Note that a string can't be accepted since zipfile has no way to
guess what encoding should be used to turn it into bytes.

Patch by Victor Stinner.


git-svn-id: http://svn.python.org/projects/python/branches/py3k@87430 6015fed2-1504-0410-9fe1-9d1591cc4771